+++ cmake/nfxGraphicsSources.cmake
@@ -6,6 +6,7 @@ set(NFX_GRAPHICS_SOURCES
     src/gl/material/Material.cpp
     src/gl/mesh/Mesh.cpp
     src/gl/mesh/Primitive.cpp
+    src/gl/pipeline/passes/GeometryPass.cpp
     src/gl/pipeline/Renderer.cpp
     src/gl/pipeline/RenderState.cpp
     src/gl/pipeline/RenderTarget.cpp